refactor(plugins): converge styles to index.cssr + per-component cssr
- session: styles -> components/archive-panel.cssr.ts, mounted via
  useMountStyle; apply no longer mounts styles
- rightclick: menu + toast styles merged into styles/index.cssr.ts
- placeholder: styles/placeholder.cssr.ts -> styles/index.cssr.ts
- dsh-tauri-ui: turn-navigation -> styles/index.cssr.ts
- pet: split into components/pet-settings.cssr.ts + styles/index.cssr.ts
  (sidebar-entry DOM patch styles)
