+++ packages/dsh-tauri-pet/src/client/components/pet-settings.tsx
@@ -1,6 +1,6 @@
 import type { ChangeEvent, ReactElement } from 'react'
 import type { PetListItem, PetSettingsProps, PresetDownloadProgress, PresetPetItem } from '../types'
-import { ArrowDownToLine, Icon, Plus } from 'dsh-tauri-ui/client'
+import { ArrowDownToLine, Icon, Plus, useMountStyle } from 'dsh-tauri-ui/client'
 import { useCallback, useEffect, useRef, useState, useSyncExternalStore } from 'react'
 import { BUILTIN_PET_ID, PET_DEFAULT_SIZE, PET_SIZE_MAX, PET_SIZE_MIN, PET_SIZE_STEP } from '../constants'
 import { text, usePetLocale } from '../locales'
@@ -19,6 +19,7 @@ import {
 } from '../service/pet'
 import { beginPetStatusFetch, commitPetStatusFetch, getPetUiSnapshot, setPetStatus, subscribePetUi } from '../store'
 import { progressPercent, resolvePresetCardAction } from '../utils/preset-card'
+import petSettingsStyle from './pet-settings.cssr'
 
 /** 预设宠物下载轮询间隔（ms）。 */
 const PRESET_DOWNLOAD_POLL_MS = 400
@@ -121,6 +122,7 @@ export function presetCardAction(
 }
 
 export function PetSettings(props: PetSettingsProps): ReactElement {
+  useMountStyle(petSettingsStyle, 'dsh-tauri-pet-settings-styles')
   usePetLocale()
   const { status } = useSyncExternalStore(subscribePetUi, getPetUiSnapshot, getPetUiSnapshot)
   const [tab, setTab] = useState<'pets' | 'codex'>('pets')
